+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# v3.2.3
+* Fixed: Not adding all of the geoshape fields to index for existing elements
+
 # v3.2.2
 * Fixed: Improved support for multithreaded clients with InMemoryGraph